Made it possible to send message from another thread
For that, I changed the uid counter from `Int` to `MVar Int` (which can
be shared and used between multiple threads).

Now, on top of the of `sendMessage` (`ChannelId -> T.Text -> Slack s ()`,
there is `sendMessageLater` (`Slack s (ChannelId -> T.Text -> IO ())`),
which returns an IO action that can be passed to another thread can be
used to send a message later.