Be more flexible parsing inconsistent Time values
Slack's API sometimes returns strings and sometimes returns integers for
times. This was causing the `channel_rename` feature to fail, since it
was returning the channels `created` attribute as a string, and we were
only parsing integer values. This commit fixes that by adding a second
function head to parse string values.

Fixes: https://github.com/mpickering/slack-api/issues/34