Correct FileChange Event
The "file\_change" event did not conform to the Slack API. It was attempting to read an entire "file" object, as though this was a "file\_created" event. Making matters a bit more complicated the Slack API itself does not correctly document all of the fields on the event.

Due to the somewhat large number of fields, and in an attempt to keep with general idioms for the `Event` GADT, the `FileChange` event now wraps a `FileChangeInfo` value. The primary advantage here is that we get record functions for accessing the members. In the future we may want to consider some more significant changes.